Fixed GameCamera

+++ Editor/Source/Screen/GameViewportScreen.cpp
@@ -25,25 +25,27 @@ namespace Editor
 		{
 			return;
 		}
-		
+
 		auto* renderTarget = &gameCameraComponent->GetRenderTarget();
 
 		if ((m_ViewportSize.x > 0.0f && m_ViewportSize.y > 0.0f) &&
-			(m_ViewportSize.x != renderTarget->GetInfo().Width ||
-			m_ViewportSize.y != renderTarget->GetInfo().Height))
+		    (m_ViewportSize.x != renderTarget->GetInfo().Width ||
+		     m_ViewportSize.y != renderTarget->GetInfo().Height))
 		{
 			gameCameraComponent->SetSize(m_ViewportSize.x, m_ViewportSize.y);
 			renderTarget = &gameCameraComponent->GetRenderTarget();
 		}
-		
+
 		ImGui::Begin(GetNameAndIDLabel());
 
 		ImVec2 displaySize = ImGui::GetContentRegionAvail();
-		m_ViewportSize.x = displaySize.x;
-		m_ViewportSize.y = displaySize.y;
+		m_ViewportSize.x   = (uint32_t)displaySize.x;
+		m_ViewportSize.y   = (uint32_t)displaySize.y;
 
-		ImGui::Image(&renderTarget->GetFrame(), ImVec2(m_ViewportSize.x, m_ViewportSize.y));
-		
+		ImGui::Image(&renderTarget->GetFrame(),
+		             ImVec2((float)m_ViewportSize.x,
+		                    (float)m_ViewportSize.y));
+
 		ImGui::End();
 	}
 }
